Try to fix bugs which prevent compiling the nigly builds.
I have no clang but it is a shame that gcc did not find the problems.

- the clang stuff shall be fixed
- most of the icc stuff shall be fixed but not the shared pointer stuff yet
- the mingw stuff could work now



82% tests passed, 17 tests failed out of 95

Total Test time (real) =  61.37 sec

The following tests FAILED:
	  5 - Singlefield_Acoustics_PmlVarDensity2d (Failed)
	 14 - Singlefield_Acoustics_Mortar2d (Failed)
	 27 - Singlefield_Heat_HeatMortar (Failed)
	 44 - Singlefield_Magnetics_Coil2dMortar (Failed)
	 51 - Singlefield_Mechanics_PreStress3dEV (Failed)
	 60 - Singlefield_Mechanics_dampedEV2D (Failed)
	 61 - Singlefield_Mechanics_BlochMode2D (Failed)
	 62 - Singlefield_Mechanics_BlochMode3D (Failed)
	 68 - Coupledfield_MechAcou_RingHarm2d (Failed)
	 78 - Coupledfield_MagMech_PermaMSRestart (Failed)
	 79 - Coupledfield_MagMechAcou_PlateAboveCoilFluid2D (Failed)
	 86 - Optimization_Homogenization_homogenize_only (Failed)
	 88 - Optimization_Dynamic_eigenfrequency (Failed)
	 89 - Optimization_Dynamic_bloch2d (Failed)
	 92 - Optimization_FMO_mech_coef_ev (Failed)
	 93 - Optimization_ParamMat_TopMat_DensityTimesRotatedTransversalIsotropic (Failed)
	 94 - Optimization_ParamMat_TopMat_DensityTimesRotatedInterpolatedTensor (Failed)

+++ source/Driver/EigenFrequencyDriver.cc
@@ -371,7 +371,7 @@ namespace CoupledField {
           // time is step.nr
           int total = eigenFreqs->GetSize() * wvs;
           int digs =  boost::lexical_cast<string>(total).size();
-          double sig = std::pow(10, -digs); // 1e-2 -> 10 ^ -2
+          double sig = std::pow((float) 10.0, -digs); // 1e-2 -> 10 ^ -2 -> a compiler complained with simply 10
           save_value = stepNum + (w * wvs + fi + 1) * sig; // +1 for one based
 
           LOG_DBG3(efd) << "SR total=" << total << " digs=" << digs << " sig=" << sig << " count=" << (w * wvs + fi + 1);
